[platform_device_loader] isolate remove device related operations
