+++ libparsec/crates/platform_device_loader/src/lib.rs
@@ -6,7 +6,7 @@ pub use device::*;
 use libparsec_platform_async::stream::StreamExt;
 use libparsec_platform_filesystem::{
     list_files, load_file, remove_file, save_content, ListFilesError, LoadFileError,
-    RemoveFileError, SaveContentError,
+    SaveContentError,
 };
 pub use strategy::*;
 
@@ -115,40 +115,6 @@ pub async fn save_device(
 
 pub fn is_keyring_available() -> bool {
     platform::is_keyring_available()
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum RemoveDeviceError {
-    #[error("Device storage is not available")]
-    StorageNotAvailable,
-    #[error("Device not found")]
-    NotFound,
-    #[error(transparent)]
-    Internal(#[from] anyhow::Error),
-}
-
-impl From<RemoveFileError> for RemoveDeviceError {
-    fn from(value: RemoveFileError) -> Self {
-        match value {
-            RemoveFileError::NotFound => RemoveDeviceError::NotFound,
-            RemoveFileError::StorageNotAvailable => RemoveDeviceError::StorageNotAvailable,
-            RemoveFileError::InvalidParent
-            | RemoveFileError::InvalidPath
-            | RemoveFileError::Internal(_) => RemoveDeviceError::Internal(value.into()),
-        }
-    }
-}
-
-pub async fn remove_device(
-    #[cfg_attr(not(feature = "test-with-testbed"), expect(unused_variables))] config_dir: &Path,
-    device_path: &Path,
-) -> Result<(), RemoveDeviceError> {
-    #[cfg(feature = "test-with-testbed")]
-    if let Some(result) = testbed::maybe_remove_device(config_dir, device_path) {
-        return result;
-    }
-
-    Ok(remove_file(device_path).await?)
 }
 
 fn encrypt_device(device: &LocalDevice, key: &SecretKey) -> Bytes {
